📷 Update snapshot test for TokenList
